pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "asyncpg>=0.30.0",
    "discord-py>=2.5.2",
    "metaapi-cloud-sdk>=28.0.7",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "tgcrypto>=1.2.5",
]
